refactor(headers): Phase 0 spike — relocate 7 low-consumer headers + deprecation shims (GH #1280, CoolProp-dfw)
Validates the git-mv + include-rewrite + checked-in-shim + preflight
mechanics ahead of the bulk header reorg into include/CoolProp/.

Public headers -> include/CoolProp/ tiers (deprecation shim left at old flat path):
  finite_diff.h     -> CoolProp/numerics/finite_diff.h
  ODEIntegrators.h  -> CoolProp/numerics/ODEIntegrators.h
  CPmsgpack.h       -> CoolProp/detail/msgpack.h        (renamed)
  Ice.h             -> CoolProp/fluids/Ice.h

Internal/test/dev headers -> src/ (ejected; not public, no shim; resolve
via the src/ search path):
  SpeedTest.h          -> src/SpeedTest.h
  Tests.h              -> src/Tests/Tests.h
  qmass_conversions.h  -> src/qmass_conversions.h

Each flat shim emits a #pragma message deprecation (portable across
GCC/Clang/MSVC), silenceable via COOLPROP_NO_DEPRECATED_HEADER_WARNINGS,
and forwards to the canonical CoolProp/<tier>/ path. Internal consumers
rewritten to the canonical paths so the in-tree build stays warning-free.
