+++ dev/codelite/coolprop.project
@@ -138,15 +138,15 @@
     <File Name="../../include/gitrevision.h"/>
     <File Name="../../include/Helmholtz.h"/>
     <File Name="../../include/HumidAirProp.h"/>
-    <File Name="../../include/Ice.h"/>
+    <File Name="../../include/CoolProp/fluids/Ice.h"/>
     <File Name="../../include/IncompressibleFluid.h"/>
     <File Name="../../include/MatrixMath.h"/>
     <File Name="../../include/PlatformDetermination.h"/>
     <File Name="../../include/PolyMath.h"/>
     <File Name="../../include/Solvers.h"/>
-    <File Name="../../include/SpeedTest.h"/>
+    <File Name="../../src/SpeedTest.h"/>
     <File Name="../../include/TestObjects.h"/>
-    <File Name="../../include/Tests.h"/>
+    <File Name="../../src/Tests/Tests.h"/>
     <File Name="../../include/Ancillaries.h"/>
     <File Name="../../include/PhaseEnvelope.h"/>
     <File Name="../../include/Configuration.h"/>
