+++ include/Ice.h
@@ -1,14 +1,6 @@
-#ifndef ICE_H
-#define ICE_H
-
-double psub_Ice(double T);
-double g_Ice(double T, double p);
-double dg_dp_Ice(double T, double p);
-double dg2_dp2_Ice(double T, double p);
-double IsothermCompress_Ice(double T, double p);
-double dg_dT_Ice(double T, double p);
-double h_Ice(double T, double p);
-double s_Ice(double T, double p);
-double rho_Ice(double T, double p);
-
+// DEPRECATED forwarding shim (GH #1280). Remove at v9.
+// The canonical location is <CoolProp/fluids/Ice.h>.
+#ifndef COOLPROP_NO_DEPRECATED_HEADER_WARNINGS
+#  pragma message("Ice.h is deprecated; include \"CoolProp/fluids/Ice.h\". Define COOLPROP_NO_DEPRECATED_HEADER_WARNINGS to silence.")
 #endif
+#include "CoolProp/fluids/Ice.h"
